backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic==2.5.0
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0
github3.py==4.0.1
packaging==23.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3 